+++ Tests/GRDBTests/FTS5TableBuilderTests.swift
@@ -36,7 +36,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .ascii()
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='ascii')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''ascii''')")
         }
     }
 
@@ -47,7 +47,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .ascii(separators: ["X"])
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='ascii separators ''X''')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''ascii'' ''separators'' ''X''')")
         }
     }
 
@@ -58,7 +58,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .ascii(tokenCharacters: Set(".-"))
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='ascii tokenchars ''-.''')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''ascii'' ''tokenchars'' ''-.''')")
         }
     }
 
@@ -69,7 +69,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .porter()
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='porter')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''porter''')")
         }
     }
 
@@ -80,7 +80,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .porter(wrapping: .ascii())
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='porter ascii')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''porter'' ''ascii''')")
         }
     }
 
@@ -91,7 +91,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .porter(wrapping: .unicode61())
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='porter unicode61')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''porter'' ''unicode61''')")
         }
     }
 
@@ -102,7 +102,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61()
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61''')")
         }
     }
 
@@ -113,7 +113,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61(diacritics: .keep)
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61 remove_diacritics 0')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61'' ''remove_diacritics'' ''0''')")
         }
     }
 
@@ -125,7 +125,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61(diacritics: .remove)
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61 remove_diacritics 2')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61'' ''remove_diacritics'' ''2''')")
         }
     }
     #elseif !GRDBCIPHER
@@ -140,7 +140,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61(diacritics: .remove)
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61 remove_diacritics 2')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61'' ''remove_diacritics'' ''2''')")
         }
     }
     #endif
@@ -152,7 +152,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61(separators: ["X"])
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61 separators ''X''')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61'' ''separators'' ''X''')")
         }
     }
 
@@ -163,7 +163,7 @@ class FTS5TableBuilderTests: GRDBTestCase {
                 t.tokenizer = .unicode61(tokenCharacters: Set(".-"))
                 t.column("content")
             }
-            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='unicode61 tokenchars ''-.''')")
+            assertDidExecute(sql: "CREATE VIRTUAL TABLE \"documents\" USING fts5(content, tokenize='''unicode61'' ''tokenchars'' ''-.''')")
         }
     }